New Version
Former-commit-id: 14a0f12cf36180e6cd7df31be85ee99432334cdd

+++ src/librapid/array/multiarray_indexing.cpp
@@ -2,7 +2,7 @@
 
 namespace librapid
 {
-	const Array Array::subscript(size_t index) const
+	const Array Array::subscript(int64_t index) const
 	{
 		if (index >= m_extent[0])
 		{
